extract ConnectButton
