+++ src/components/common/Button.tsx
@@ -23,7 +23,7 @@ const variantClasses: Record<NonNullable<ButtonProps["variant"]>, string> = {
 const sizeClasses: Record<NonNullable<ButtonProps["size"]>, string> = {
   sm: "h-8 px-3 text-sm",
   md: "h-10 px-4 text-sm",
-  lg: "h-12 px-6 text-base",
+  lg: "h-14 px-6 text-base",
 };
 
 export default function Button(allProps: ButtonProps): JSX.Element {
